FLEX_NEXT: major decoder rewrite per ARIB STD-43A
Sync/demod:
- S2 C/inv.C detection with symbol buffer and replay for boundary correction
- Phase mapping: bit_b to PhaseC at 1600 baud, PhaseB at 3200 baud
- Subframe decode for multiple transmission (num_tx 2/3/4)

Frame parsing:
- FIW: num_tx, td_collapse; BIW101 A-type and system message vector
- Address types: N/T/O/I/R with network payload, operator category
- Tone-only pre-scan via vector checksum; long address guard
- FLEX_PAGETYPE_TONE_ONLY for addresses without vector word
- FLEX_PAGETYPE_SHORT_MESSAGE rename (was TONE) for V=010

Header word extraction:
- Type-dependent bit layout: HEX 12-bit K (C/F/N shifted vs alpha)
- HEX R/M from hdr2; B at bits 4-7, S at bits 13-20

Fragment reassembly:
- 16 slots, 64-frame timeout, keyed by (capcode, type, msg_n)
- F sequence tracking: mod 3 cycle with mismatch detection
- R=1 evicts stale slot; R/M carried to continuations and reassembled
- Per-fragment metadata: sequence, index, word/char counts

Signatures and checksums:
- Alpha: 7-bit signature excluding ETX, accumulated across fragments
- HEX: 8-bit signature from data words only
- K checksum per fragment; .K- and .SIGN- flags on failure

Word-level deduplication:
- 32-slot cache, raw 21-bit words + BCH status
- Retransmissions merged at word level; DUP/DUP+ flags

JSON output (--json):
- All message types: ALN, NUM/SNUM/NNUM, HEX, SEC, SMSG, TON, INS, NET
- BIW system info: biw_sysid, biw_date, biw_time, biw_sysinfo, biw_countrycode
- Human-readable: msg_type, instruction_type, sec_subtype, smsg_sub_type
- Group messages: group_capcodes array, group_slot
- Fragment info: frag_seq, frag_index, frag_seq_error, total_fragments, total_chars
- Extra fields: blocking, source_code, special_format, opr_category, biw_position

flextime (OTA timestamp):
- BIW date/time/timezone with per-component age tracking (frames and seconds)
- Timezone lookup table: 32 entries with fractional offsets per FLEX spec
- Extended seconds: S5-S3 + S2-S0 combined for 0.9375s precision
- Wrap-aware expiry after one full cycle (1920 frames)
- precision_seconds field (7.5 or 0.9375)

Output format:
- Frag flags: F.C.flag.Nnn.Rn[.M][.K-][.SIGN-][.DUP][.DUP+]
- Group capcodes in capcode field (space-separated)
- Atomic level-0 output prevents debug interleaving

Bug fixes:
- Group message K/signature: len always decremented for short addresses
- Instruction vector (V=001): handler moved before hdr/len calculation
- Instruction vector pre-scan: bypass nibble checksum for type=1
- F sequence: advance expected_f only on continuation, not initial

Test suite:
- FLEX_NEXT 3200/4FSK sample from NYC 929.6125 MHz IQ capture
- Tests 4-phase decode, long addresses, fragment reassembly, checksums
- gen-ng FLEX_NEXT roundtrip test (gen -> decode -> verify)

+++ gen.c
@@ -283,11 +283,11 @@ static void output_sound(unsigned int sample_rate, const char *ifname)
 		perror("ioctl: SNDCTL_DSP_SPEED");
 		exit (10);
 	}
-        if ((10*abs(sndparam-sample_rate)) > sample_rate) {
+        if ((10*(unsigned)abs((int)((unsigned)sndparam-sample_rate))) > sample_rate) {
 		perror("ioctl: SNDCTL_DSP_SPEED");
 		exit (10);
 	}
-        if (sndparam != sample_rate) {
+        if ((unsigned)sndparam != sample_rate) {
                 fprintf(stderr, "Warning: Sampling rate is %u, "
                         "requested %u\n", sndparam, sample_rate);
         }
